+++ eligibilityd/Eligibility+Util.m
@@ -8,7 +8,7 @@
 #import "Eligibility+Util.h"
 #import "EligibilityLog.h"
 
-NSString * _Nullable eligibility_input_to_str(uint64_t type) {
+NSString * _Nullable eligibility_input_to_str(NSUInteger type) {
     switch (type) {
         case 0: return @"OS_ELIGIBILITY_INPUT_INVALID";
         case 1: return @"OS_ELIGIBILITY_INPUT_COUNTRY_LOCATION";
@@ -22,12 +22,12 @@ NSString * _Nullable eligibility_input_to_str(uint64_t type) {
         case 9: return @"OS_ELIGIBILITY_INPUT_GREYMATTER_ON_QUEUE";
         case 10: return @"OS_ELIGIBILITY_INPUT_SIRI_LANGUAGE";
         default:
-            os_log_error(eligibility_log(), "%s: Unsupported input type: %llu", "eligibility_input_to_str", type);
+            os_log_error(eligibility_log(), "%s: Unsupported input type: %llu", "eligibility_input_to_str", (unsigned long long)type);
             return nil;
     }
 }
 
-NSString * _Nullable eligibility_input_status_to_str(uint64_t status) {
+NSString * _Nullable eligibility_input_status_to_str(NSUInteger status) {
     switch (status) {
         case 0: return @"OS_ELIGIBILITY_INPUT_STATUS_NONE";
         case 1: return @"OS_ELIGIBILITY_INPUT_STATUS_NOT_SET";
@@ -38,7 +38,7 @@ NSString * _Nullable eligibility_input_status_to_str(uint64_t status) {
         case 6: return @"OS_ELIGIBILITY_INPUT_STATUS_TOKEN_EXPIRED";
         case 7: return @"OS_ELIGIBILITY_INPUT_STATUS_NO_ACCOUNT";
         default:
-            os_log_error(eligibility_log(), "%s: Unsupported input status: %llu", "eligibility_input_status_to_str", status);
+            os_log_error(eligibility_log(), "%s: Unsupported input status: %llu", "eligibility_input_status_to_str", (unsigned long long) status);
             break;
     }
     return nil;
